Modificaciones en condiciones creación usuarios

+++ src/main/java/com/hortalsoft/users/domain/repository/UserRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<UserEntity, Integer> {
         Optional<UserEntity> findByEmailAndIdNumber(String email, long idNumber);
         Optional<UserEntity> findByEmail(String email);
-        Optional<UserEntity> findByIdNumber(Integer idNumber);
+        Optional<UserEntity> findByIdNumber(Long idNumber);
 
 }
